frontend: Make roots first-class objects.
