+++ packages/frontend-web/src/app/components/SplashRootStory.tsx
